backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
openai
PyMuPDF
httpx
jsonschema
requests
beautifulsoup4
pdfplumber
fast-json-repair
tenacity
